+++ src/scrape/allRegions.ts
@@ -1,0 +1,78 @@
+export const allRegions = [
+  'Asia',
+  'Central America',
+  'Albania',
+  'Austria',
+  'Bavaria',
+  'Friuli Venezia Giulia',
+  'Lombardia',
+  'Piemonte',
+  "Provence-Alpes-Cote d'Azur",
+  'Slovenia',
+  'Switzerland',
+  'Trentino-Alto Adige',
+  "Valle d'Aosta",
+  'Andorra',
+  'Bulgaria',
+  'Croatia',
+  'France',
+  'Germany',
+  'Greece',
+  'Hungary',
+  'Iceland',
+  'Ireland',
+  'Islas Canarias',
+  'Italy',
+  'Lithuania',
+  'Macedonia',
+  'Montenegro',
+  'Poland',
+  'Portugal',
+  'Romania',
+  'Slovakia',
+  'Spain',
+  'Turkey',
+  'United Kingdom',
+  'Middle East',
+  'Canada',
+  'Mexico',
+  'Northeast',
+  'Pacific Northwest',
+  'Alaska',
+  'Arizona',
+  'Arkansas',
+  'California',
+  'Utah',
+  'Colorado',
+  'Connecticut',
+  'Georgia',
+  'Hawaii',
+  'Idaho',
+  'Maine',
+  'Massachusetts',
+  'Montana',
+  'Nevada',
+  'New Mexico',
+  'New York',
+  'North Carolina',
+  'Oregon',
+  'South Carolina',
+  'South Dakota',
+  'Texas',
+  'Utah',
+  'Virginia',
+  'Washington',
+  'West Desert',
+  'Wyoming',
+  'Pacific',
+  'South America',
+  'Argentina',
+  'Bolivia',
+  'Brazil',
+  'Chile',
+  'Colombia',
+  'Ecuador',
+  'Peru',
+  'Playa Montezuma',
+  'Venezuela',
+];
